+++ docs/Whats-New.md
@@ -359,6 +359,7 @@ New:
 - Jumpjet Tilts While Moving (by CrimRecya)
 - [Spawned aircraft facing to match turret toggle](New-or-Enhanced-Logics.md#aircraft-spawner-customizations) (by Starkku)
 - Removed dependency on `blowfish.dll` (by ZivDero)
+- Restore turret recoil effect (by CrimRecya)
 
 Vanilla fixes:
 - Prevent the units with locomotors that cause problems from entering the tank bunker (by TaranDahl)